Bug fix
 * When creating 1 Mbyte images, the *PROM partition address
   resets after 0x10000

+++ ROM_image_generator.c
@@ -3,8 +3,8 @@
 #include <string.h>
 
 typedef struct partitions_s {
-  uint16_t start;
-  uint16_t end;
+  uint32_t start;
+  uint32_t end;
   char *name;
 } partitions_t;
